refactor: replace SetCurrentVaultIdProvider with WriteStorageProvider in AppProviders, streamline vault state management, and update related components for improved clarity and functionality
