+++ clients/desktop/src/vault/state/currentVaultId.tsx
@@ -4,7 +4,6 @@ import {
   getInitialVaultId,
   useCurrentVaultIdCorrector,
 } from '@core/ui/vault/state/currentVaultId'
-import { SetCurrentVaultIdProvider as CoreSetCurrentVaultIdProvider } from '@core/ui/vault/state/setCurrentVaultId'
 import { useVaults } from '@core/ui/vault/state/vaults'
 import { ChildrenProp } from '@lib/ui/props'
 import { useStateCorrector } from '@lib/ui/state/useStateCorrector'
@@ -15,7 +14,7 @@ import {
   usePersistentState,
 } from '../../state/persistentState'
 
-const useCurrentVaultId = (): [
+export const useCurrentVaultId = (): [
   CurrentVaultId,
   Dispatch<SetStateAction<CurrentVaultId>>,
 ] => {
@@ -39,15 +38,5 @@ export const CurrentVaultIdProvider = ({ children }: ChildrenProp) => {
     <CoreCurrentVaultIdProvider value={currentVaultId}>
       {children}
     </CoreCurrentVaultIdProvider>
-  )
-}
-
-export const SetCurrentVaultIdProvider = ({ children }: ChildrenProp) => {
-  const [, setCurrentVaultId] = useCurrentVaultId()
-
-  return (
-    <CoreSetCurrentVaultIdProvider value={setCurrentVaultId}>
-      {children}
-    </CoreSetCurrentVaultIdProvider>
   )
 }
