+++ clients/desktop/src/vaults/components/VaultListItem.tsx
@@ -1,6 +1,6 @@
 import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
+import { useSetCurrentVaultIdMutation } from '@core/ui/vault/mutations/useSetCurrentVaultIdMutation'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
-import { useSetCurrentVaultId } from '@core/ui/vault/state/setCurrentVaultId'
 import { getVaultId } from '@core/ui/vault/Vault'
 
 import { ListItem } from '../../lib/ui/list/item/ListItem'
@@ -12,7 +12,7 @@ type VaultListItemProps = {
 
 export const VaultListItem = ({ isDraggable }: VaultListItemProps) => {
   const navigate = useCoreNavigate()
-  const setCurrentVaultId = useSetCurrentVaultId()
+  const { mutate: setCurrentVaultId } = useSetCurrentVaultIdMutation()
   const vault = useCurrentVault()
 
   return (
