+++ clients/extension/src/state/storage/write.tsx
@@ -5,9 +5,11 @@ import {
 import { ChildrenProp } from '@lib/ui/props'
 
 import { setFiatCurrency } from '../../preferences/fiatCurrency'
+import { setCurrentVaultId } from '../../vault/state/currentVaultId'
 
 const writeStorage: CoreWriteStorage = {
   setFiatCurrency,
+  setCurrentVaultId,
 }
 
 export const WriteStorageProvider = ({ children }: ChildrenProp) => {
