--- clients/extension/src/vault/state/setCurrentVaultIdProvider.tsx
@@ -1,12 +1,0 @@
-import { SetCurrentVaultIdProvider as SetCurrentVaultIdProviderBase } from '@core/ui/vault/state/setCurrentVaultId'
-import { ChildrenProp } from '@lib/ui/props'
-
-import { setCurrentVaultId } from './currentVaultId'
-
-export const SetCurrentVaultIdProvider = ({ children }: ChildrenProp) => {
-  return (
-    <SetCurrentVaultIdProviderBase value={setCurrentVaultId}>
-      {children}
-    </SetCurrentVaultIdProviderBase>
-  )
-}
