+++ core/ui/vault/mutations/useSetCurrentVaultIdMutation.ts
@@ -2,15 +2,15 @@ import { currentVaultIdQueryKey } from '@core/ui/query/keys'
 import { useInvalidateQueries } from '@lib/ui/query/hooks/useInvalidateQueries'
 import { useMutation, UseMutationOptions } from '@tanstack/react-query'
 
+import { useCoreWriteStorage } from '../../state/storage/write'
 import { CurrentVaultId } from '../state/currentVaultId'
-import { useSetCurrentVaultId } from '../state/setCurrentVaultId'
 
 export const useSetCurrentVaultIdMutation = (
   options?: UseMutationOptions<any, any, CurrentVaultId, unknown>
 ) => {
   const invalidateQueries = useInvalidateQueries()
 
-  const setCurrentVaultId = useSetCurrentVaultId()
+  const { setCurrentVaultId } = useCoreWriteStorage()
 
   const mutationFn = async (value: CurrentVaultId) => {
     await setCurrentVaultId(value)
